+++ spec/lib/tasks/export_task_spec.rb
@@ -6,14 +6,14 @@ RSpec.describe 'Export Rake task' do
   it 'halts when the API key is not set' do
     allow(global_config).to receive(:api_token).and_return(nil)
 
-    expect(export_executor).to raise_error(SystemExit, /API token is not set/i)
+    expect { Rake::Task['lokalise_rails:export'].execute }.to raise_error(SystemExit, /API token is not set/i)
     expect(global_config).to have_received(:api_token)
   end
 
   it 'halts when the project ID is not set' do
     allow(global_config).to receive(:project_id).and_return(nil)
 
-    expect(export_executor).to raise_error(SystemExit, /ID is not set/i)
+    expect { Rake::Task['lokalise_rails:export'].execute }.to raise_error(SystemExit, /ID is not set/i)
 
     expect(global_config).to have_received(:project_id)
   end
@@ -22,11 +22,11 @@ RSpec.describe 'Export Rake task' do
     let(:filename_ru) { 'ru.yml' }
     let(:path_ru) { "#{Rails.root}/config/locales/#{filename_ru}" }
 
-    before :all do
+    before do
       add_translation_files! with_ru: true
     end
 
-    after :all do
+    after do
       rm_translation_files
     end
 
@@ -34,7 +34,7 @@ RSpec.describe 'Export Rake task' do
       it 'is callable' do
         allow_project_id global_config, ENV['LOKALISE_PROJECT_ID'] do
           VCR.use_cassette('upload_files') do
-            expect(export_executor).to output(/complete!/).to_stdout
+            expect { Rake::Task['lokalise_rails:export'].execute }.to output(/complete!/).to_stdout
           end
         end
       end
@@ -42,7 +42,7 @@ RSpec.describe 'Export Rake task' do
       it 're-raises export errors' do
         allow_project_id global_config, '542886116159f798720dc4.94769464' do
           VCR.use_cassette('upload_files_error') do
-            expect(export_executor).to raise_error(SystemExit, /Unknown `lang_iso`/)
+            expect { Rake::Task['lokalise_rails:export'].execute }.to raise_error(SystemExit, /Unknown `lang_iso`/)
           end
         end
       end